requirements.txt
# Core dependencies
watchdog
dateparser
deep-translator
openai
python-dotenv
requests
pandas
rich
docxtpl
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
notion-client
pyyaml
fastapi
boto3
awscli
fastapi 
uvicorn 
jinja2
python-multipart

# Development tools (optional, comment out if not needed)
#black
